requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
annotated-types=0.7.0=pypi_0
anyio=4.9.0=pypi_0
appdirs=1.4.4=pypi_0
argcomplete=3.6.2=pypi_0
asgiref=3.8.1=pypi_0
attr=0.3.1=pypi_0
attrs=25.3.0=pypi_0
azure-core=1.34.0=pypi_0
azure-storage-blob=12.25.1=pypi_0
black=25.1.0=pypi_0
bleach=5.0.1=pypi_0
boto=2.49.0=pypi_0
boto3=1.38.17=pypi_0
botocore=1.38.17=pypi_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.2.25=haa95532_0
cachetools=5.5.2=pypi_0
certifi=2025.4.26=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.2=pypi_0
click=8.2.0=pypi_0
colorama=0.4.6=pypi_0
contourpy=1.3.2=pypi_0
cryptography=44.0.3=pypi_0
cycler=0.12.1=pypi_0
datamodel-code-generator=0.26.1=pypi_0
defusedxml=0.7.1=pypi_0
deprecated=1.2.18=pypi_0
distro=1.9.0=pypi_0
django=5.1.9=pypi_0
django-annoying=0.10.6=pypi_0
django-cors-headers=3.6.0=pypi_0
django-csp=3.7=pypi_0
django-debug-toolbar=3.2.1=pypi_0
django-environ=0.10.0=pypi_0
django-extensions=3.2.3=pypi_0
django-filter=24.3=pypi_0
django-migration-linter=5.2.0=pypi_0
django-model-utils=4.1.1=pypi_0
django-ranged-fileresponse=0.1.2=pypi_0
django-rq=2.10.3=pypi_0
django-storages=1.12.3=pypi_0
django-user-agents=0.4.0=pypi_0
djangorestframework=3.15.2=pypi_0
djangorestframework-simplejwt=5.4.0=pypi_0
dnspython=2.7.0=pypi_0
drf-dynamic-fields=0.3.0=pypi_0
drf-flex-fields=0.9.5=pypi_0
drf-generators=0.3.0=pypi_0
email-validator=2.2.0=pypi_0
expat=2.7.1=h8ddb27b_0
expiringdict=1.2.2=pypi_0
faker=37.3.0=pypi_0
filelock=3.18.0=pypi_0
fonttools=4.58.0=pypi_0
fsspec=2025.3.2=pypi_0
genson=1.3.0=pypi_0
google-api-core=2.24.2=pypi_0
google-auth=2.40.1=pypi_0
google-cloud-appengine-logging=1.6.1=pypi_0
google-cloud-audit-log=0.3.2=pypi_0
google-cloud-core=2.4.3=pypi_0
google-cloud-logging=3.12.1=pypi_0
google-cloud-storage=2.19.0=pypi_0
google-crc32c=1.7.1=pypi_0
google-resumable-media=2.7.2=pypi_0
googleapis-common-protos=1.70.0=pypi_0
grpc-google-iam-v1=0.14.2=pypi_0
grpcio=1.72.0rc1=pypi_0
grpcio-status=1.72.0rc1=pypi_0
h11=0.16.0=pypi_0
httpcore=1.0.9=pypi_0
httpx=0.28.1=pypi_0
humansignal-drf-yasg=1.21.10.post1=pypi_0
idna=3.10=pypi_0
ijson=3.4.0=pypi_0
importlib-metadata=8.6.1=pypi_0
inflect=5.6.2=pypi_0
inflection=0.5.1=pypi_0
isodate=0.7.2=pypi_0
isort=5.13.2=pypi_0
jinja2=3.1.6=pypi_0
jiter=0.9.0=pypi_0
jmespath=1.0.1=pypi_0
joblib=1.5.0=pypi_0
jsf=0.11.2=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2025.4.1=pypi_0
kiwisolver=1.4.8=pypi_0
label-studio=1.18.0=pypi_0
label-studio-sdk=1.0.12=pypi_0
launchdarkly-server-sdk=8.2.1=pypi_0
libffi=3.4.4=hd77b12b_1
lockfile=0.12.2=pypi_0
lxml=5.4.0=pypi_0
lxml-html-clean=0.4.2=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=pypi_0
matplotlib=3.10.3=pypi_0
mdurl=0.1.2=pypi_0
mpmath=1.3.0=pypi_0
mypy-extensions=1.1.0=pypi_0
networkx=3.4.2=pypi_0
nltk=3.9.1=pypi_0
numpy=1.26.4=pypi_0
openai=1.78.1=pypi_0
opencv-python=4.11.0.86=pypi_0
openssl=3.0.16=h3f729d1_0
opentelemetry-api=1.33.0=pypi_0
ordered-set=4.0.2=pypi_0
packaging=25.0=pypi_0
pandas=2.2.3=pypi_0
pathspec=0.12.1=pypi_0
pillow=11.2.1=pypi_0
pip=25.1=pyhc872135_2
platformdirs=4.3.8=pypi_0
proto-plus=1.26.1=pypi_0
protobuf=6.31.0=pypi_0
psutil=7.0.0=pypi_0
psycopg2-binary=2.9.10=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.2=pypi_0
pyboxen=1.3.0=pypi_0
pycparser=2.22=pypi_0
pydantic=2.11.4=pypi_0
pydantic-core=2.33.2=pypi_0
pygments=2.19.1=pypi_0
pyjwt=2.10.1=pypi_0
pyparsing=3.2.3=pypi_0
pyqt5=5.15.11=pypi_0
pyqt5-qt5=5.15.2=pypi_0
pyqt5-sip=12.17.0=pypi_0
pyrfc3339=2.0.1=pypi_0
python=3.12.9=h14ffc60_0
python-dateutil=2.9.0.post0=pypi_0
python-json-logger=2.0.4=pypi_0
pytz=2022.7.1=pypi_0
pyyaml=6.0.2=pypi_0
redis=5.2.1=pypi_0
referencing=0.36.2=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
requests-file=2.1.0=pypi_0
requests-mock=1.12.1=pypi_0
rich=14.0.0=pypi_0
rpds-py=0.25.0=pypi_0
rq=1.16.2=pypi_0
rsa=4.9.1=pypi_0
rstr=3.2.2=pypi_0
rules=3.4=pypi_0
s3transfer=0.12.0=pypi_0
scipy=1.15.3=pypi_0
semver=3.0.4=pypi_0
sentry-sdk=2.28.0=pypi_0
setuptools=78.1.1=py312haa95532_0
six=1.17.0=pypi_0
smart-open=7.1.0=pypi_0
sniffio=1.3.1=pypi_0
sqlite=3.45.3=h2bbff1b_0
sqlparse=0.5.3=pypi_0
sympy=1.14.0=pypi_0
tk=8.6.14=h0416ee5_0
tldextract=5.3.0=pypi_0
toml=0.10.2=pypi_0
torch=2.7.0=pypi_0
torchvision=0.22.0=pypi_0
tqdm=4.67.1=pypi_0
typing-extensions=4.13.2=pypi_0
typing-inspection=0.4.0=pypi_0
tzdata=2025.2=pypi_0
ua-parser=1.0.1=pypi_0
ua-parser-builtins=0.18.0.post1=pypi_0
ujson=5.10.0=pypi_0
ultralytics=8.3.136=pypi_0
ultralytics-thop=2.0.14=pypi_0
uritemplate=4.1.1=pypi_0
urllib3=1.26.20=pypi_0
user-agents=2.2.0=pypi_0
vc=14.42=haa95532_5
vs2015_runtime=14.42.34433=hbfb602d_5
webencodings=0.5.1=pypi_0
wheel=0.40.0=pypi_0
wrapt=1.17.2=pypi_0
xmljson=0.2.1=pypi_0
xz=5.6.4=h4754444_1
zipp=3.21.0=pypi_0
zlib=1.2.13=h8cc25b3_1
